requirements.txt
PySimpleGUI
Pillow
PyPDF2
fpdf2
pyinstaller
pyglet